+++ sdk/nodejs/storage/bucketObject.ts
@@ -25,7 +25,7 @@ import * as utilities from "../utilities";
  * 
  * const picture = new gcp.storage.BucketObject("picture", {
  *     bucket: "image-store",
- *     source: new pulumi.asset.FileArchive("/images/nature/garden-tiger-moth.jpg"),
+ *     source: new pulumi.asset.FileAsset("/images/nature/garden-tiger-moth.jpg"),
  * });
  * ```
  *
@@ -113,7 +113,7 @@ export class BucketObject extends pulumi.CustomResource {
      * A path to the data you want to upload. Must be defined
      * if `content` is not.
      */
-    public readonly source!: pulumi.Output<pulumi.asset.Archive | undefined>;
+    public readonly source!: pulumi.Output<pulumi.asset.Asset | pulumi.asset.Archive | undefined>;
     /**
      * The [StorageClass](https://cloud.google.com/storage/docs/storage-classes) of the new bucket object.
      * Supported values include: `MULTI_REGIONAL`, `REGIONAL`, `NEARLINE`, `COLDLINE`. If not provided, this defaults to the bucket's default
@@ -239,7 +239,7 @@ export interface BucketObjectState {
      * A path to the data you want to upload. Must be defined
      * if `content` is not.
      */
-    readonly source?: pulumi.Input<pulumi.asset.Archive>;
+    readonly source?: pulumi.Input<pulumi.asset.Asset | pulumi.asset.Archive>;
     /**
      * The [StorageClass](https://cloud.google.com/storage/docs/storage-classes) of the new bucket object.
      * Supported values include: `MULTI_REGIONAL`, `REGIONAL`, `NEARLINE`, `COLDLINE`. If not provided, this defaults to the bucket's default
@@ -290,7 +290,7 @@ export interface BucketObjectArgs {
      * A path to the data you want to upload. Must be defined
      * if `content` is not.
      */
-    readonly source?: pulumi.Input<pulumi.asset.Archive>;
+    readonly source?: pulumi.Input<pulumi.asset.Asset | pulumi.asset.Archive>;
     /**
      * The [StorageClass](https://cloud.google.com/storage/docs/storage-classes) of the new bucket object.
      * Supported values include: `MULTI_REGIONAL`, `REGIONAL`, `NEARLINE`, `COLDLINE`. If not provided, this defaults to the bucket's default